requirements/framework.txt
dotenv
faiss-cpu
json5
markdown
mcp
modelscope
numpy
omegaconf
openai
pillow
sentence-transformers
